Task 0 struct

+++ 0x0E-structures_typedef/dog.h
@@ -17,7 +17,7 @@ typedef struct dog dog_t;
 struct dog
 {
 	char *name;
-	double age;
+	float age;
 	char *owner;
 };
